fix/key_gen: update the DKG procedure flow
This change makes the Contribution phase collect all proposed
commitments and acks. And the Commitment phase do the same with
a new round of DKG that having invalid members removed.
The Justification phase is aimed to allow the complained member
to convice others to recover, however this step may not be needed.