chore: switch to bsd3 license
Currently this repository is dual license BSD/MIT. We decided our licensing policy would be to use
BSD-3-Clause for all crates without the `sn_` prefix, which will use GPL3.

To achieve this:
* The `verify-licensing-info` custom action is applied to make sure licensing is applied
  consistently.
* The LICENSE-BSD file is removed
* The LICENSE-MIT file is removed
* A LICENSE file is added with the BSD-3-Clause license
* The README is updated to reference the BSD-3-Clause license
* Source files are updated to have a BSD-3-Clause copyright reference
* The Cargo manifest is updated to use a single BSD-3-Clause license

Note that the copyright notice for the source files is intentionally short; it directs to the
license file for further reading.

A couple of unrelated things were also changed:
* A clippy error was fixed to change use of a vector to a slice
* Superfluous comments were removed from the PR workflow